Add message shortcut code for same-machine communication
(ported forward from 6.12.3)
This patch adds code to shortcut the message passing mechanism
and instead share data when a data or head message is sent
to a process on the same machine (or even to the same process).
When the receiver is indeed the same process, there is no problem.
If it is "just" the same machine, we should in principle still
duplicate the data to maintain separate heaps for the separate
processes. In practice, heaps can be shared between processes now.

Also, the new code suppresses the trace event from the trace file.
A new trace event will be introduced in a following patch, left out
here to make this patch stand-alone.