Adapted MPI compilation test to newer mpich/openmpi
The MPI options are now duplicated in both LDFLAGS and LIBS,
since many of them (not all) are indeed library flags and
need to come last on the compilation command line.

The actual usage of the flags ensures anyway that they come
after the source code when linking (see DriverPipeline.hs).