Emit (text) events for new inports and connection
With this additional data, one can build a trace processing tool
which reconstructs the program's communication structure in a DAG,
useful for debugging if explicit communication is used.

The events are, for now, textual, and need to be parsed as part of the
trace processing. Dedicated events could be defined, but this version
is ok for a first prototype.