Controlled Shutdown on Error (PVM/MPI)
This patch introduces a controlled shutdown functionality
for failing machines. A failing non-main machine will send
PP_FINISH to the main machine.
The main machine will (when failing or in a regular shutdown)
send PP_FINISH to all other machines, and expect PP_FINISH answers
which are counted to check (an earlier PP_FINISH will count towards
the sum). Therefore, the invariant is that all machines exchange
PP_FINISH messages with the main machine, in any order, before the
system shuts down.
Note also that nPEs is set to 0 at the end of MP_quit, to
avoid calling shutdownParallelSystem (and MP_quit) again.

This patch is for the MPI and PVM ways, code for the parcp way is
not ready yet.