Fixed a pvm-spawn problem
PVM 3.4.6 as installed on my Mac (Ubuntu 12.04) refused
to spawn processes where the executable was specified in
an absolute path. The error also hit elsewhere, as the
success of spawning the child processes was not checked.
This patch introduces basename() for the PVM executable
and a check of the spawn result.