Mailslot way for Windows
A new way "parms" (parallel mail slots) is defined, which
implements the MP system via Windows Mailslots.

The basic implementation (MPSystem API) is done and works for
small test programs (success case only, error case "to do").

The way is defined, but no startup script is generated (this should
be eliminated eventually for Copy way and this one).
Programs have to be started with the PE count as first arg.
at the moment; this will be straightened in a subsequent commit.