Collate eventlogs to zip archive within RTS (merged/squashed)
This patch adds a utility which creates a zip file from given
file names, and uses it for the mailslots way (for now only there)
so that the main process collates all eventlog files to a
single parevents file in zip format.
This enables us to get rid of the start script in parallel Haskell
ways which run in multiple processes on one host system.
For now, this is active for the mailslots way, it should soon be
available for the copy way as well, then also under Linux/Mac.

Note that the eventlogs are not compressed, merely archived.