Collate eventlogs to zip archive within RTS
This patch adds a utility which creates a zip file from given
file names, and uses it for the mailslots way (for now only there)
to write the parevents file which contains all event logs.
This enables us to get rid of the start script in parallel Haskell
ways which run in multiple processes on one host system.
For now, this is active for the mailslots way, it should soon be
available for the copy way as well, then also under Linux/Mac.

Note that the integration has not been tested yet, this is checked
in so I can transfer it to a Windows machine.