deactivate start script for copy way, zip trace file internally
This patch deactivates the start script for the Linux implementation
of the copy way, entailing some important changes to CpComm.
o the number of OS processes is determined from arguments in MP_start
o shared memory and semaphores are created and OS processes forked
  in MP_sync (after having parsed the RTS flags)
o after shutdown, the eventlog files are archived to a parevents
  file using integrated code.

Zlib support for archiving is currently disabled, as it will require
adding suitable linker flags, and might still have a subtle bugs
(I had added it initially, but had problems with the generated files).